requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
python-dateutil
lxml
google-generativeai
python-dotenv
groq